+++ packages/core/navigation-next/docs/20-composing-your-navigation.js
@@ -71,7 +71,7 @@ As we begin composing the navigation we might as well start at the bottom - the 
 
 ${code`
 import GlobalNavigation from '@atlaskit/global-navigation';
-import AtlassianIcon from '@atlaskit/icon/glyph/atlassian';
+import AtlassianIcon from '@atlaskit/logo';
 
 const MyGlobalNavigation = () => (
   <GlobalNavigation