+++ packages/core/navigation-next/docs/examples/ui-components/GlobalItem.js
@@ -2,7 +2,7 @@
 
 import React from 'react';
 import Avatar from '@atlaskit/avatar';
-import AtlassianIcon from '@atlaskit/icon/glyph/atlassian';
+import AtlassianIcon from '@atlaskit/logo';
 import SearchIcon from '@atlaskit/icon/glyph/search';
 import HelpIcon from '@atlaskit/icon/glyph/question';
 import { colors } from '@atlaskit/theme';