+++ packages/core/navigation-next/examples/9999-getting-started-composing-the-container-navigation.js
@@ -2,7 +2,7 @@
 
 import React from 'react';
 import GlobalNavigation from '@atlaskit/global-navigation';
-import AtlassianIcon from '@atlaskit/icon/glyph/atlassian';
+import AtlassianIcon from '@atlaskit/logo';
 import ShortcutIcon from '@atlaskit/icon/glyph/shortcut';
 import { AtlassianWordmark } from '@atlaskit/logo';
 import {