+++ packages/core/navigation/src/__tests__/unit/ContainerTitleDropdown.js
@@ -4,7 +4,7 @@ import AkDropdownMenu, {
   DropdownItem,
   DropdownItemGroup,
 } from '@atlaskit/dropdown-menu';
-import AtlassianIcon from '@atlaskit/icon/glyph/atlassian';
+import AtlassianIcon from '@atlaskit/logo';
 import ContainerTitleDropdown from '../../components/js/ContainerTitleDropdown';
 import { mountWithRootTheme } from './_theme-util';