+++ website/src/pages/Package/index.js
@@ -7,7 +7,7 @@ import { Helmet } from 'react-helmet';
 import { gridSize, colors, math } from '@atlaskit/theme';
 import Button from '@atlaskit/button';
 import ExamplesIcon from '@atlaskit/icon/glyph/screen';
-import AtlassianIcon from '@atlaskit/icon/glyph/atlassian';
+import { AtlassianIcon } from '@atlaskit/logo';
 
 import Loading from '../../components/Loading';
 import Page from '../../components/Page';